Refactor helpers package as a standard go package practice
